+++ lib/serializers/registrant_api/domain.rb
@@ -3,11 +3,14 @@ module Serializers
     class Domain
       attr_reader :domain
 
-      def initialize(domain)
+      def initialize(domain, simplify: false)
         @domain = domain
+        @simplify = simplify
       end
 
-      def to_json
+      def to_json(_obj = nil)
+        return simple_object if @simplify
+
         {
           id: domain.uuid,
           name: domain.name,
@@ -48,6 +51,17 @@ module Serializers
       end
 
       private
+
+      def simple_object
+        {
+          id: domain.uuid, name: domain.name, registered_at: domain.registered_at,
+          valid_to: domain.valid_to, outzone_at: domain.outzone_at, statuses: domain.statuses,
+          registrant_verification_asked_at: domain.registrant_verification_asked_at,
+          registrar: { name: domain.registrar.name, website: domain.registrar.website },
+          registrant: { name: domain.registrant.name, id: domain.registrant.uuid,
+                        phone: domain.registrant.phone, email: domain.registrant.email }
+        }
+      end
 
       def dnssec_keys
         domain.dnskeys.map do |key|